backend/requirements.txt
Django
django-cors-headers
django-filter
djangorestframework
djangorestframework-camel-case
djangorestframework-simplejwt
markdown
mixer